refactor(schema-api): extract CatalogApi trait from SchemaApi
Extract catalog operations into a separate trait following single
responsibility principle. This creates cleaner separation between
catalog and other schema operations, continuing the modularization
of the monolithic SchemaApi trait.

Key changes:
- Create `CatalogApi` trait in `src/meta/api/src/catalog_api.rs` with 4
  catalog management methods (create, get, drop, list)
- Move catalog method implementations from `schema_api.rs` to the new
  trait file while preserving all existing functionality
- Update `SchemaApi` trait bounds to include `CatalogApi` dependency
